Content

+++ src/app/api/approve/route.ts
@@ -1,7 +1,10 @@
 import { Collection, BSON } from "mongodb";
 import clientPromise from "@/lib/mongodb";
 import { getServerSession } from "next-auth/next";
+import algoliasearch from "algoliasearch";
 import { options } from "@/app/api/auth/[...nextauth]/options";
+const client = algoliasearch("1DNU60T3R9", "ec8bbb5e8c82c33b6037352c5d39e521"); // move to local env vars
+const index = client.initIndex("pawprint");
 
 interface Manuscript {
   _id?: BSON.ObjectId;
@@ -17,9 +20,8 @@ interface Manuscript {
 }
 
 export async function POST(req: Request) {
-  const { actionType, id, title, content, path, category, img, comment } = JSON.parse(
-    await req.text()
-  );
+  const { actionType, id, title, content, path, category, img, comment } =
+    JSON.parse(await req.text());
   const session = await getServerSession(options);
 
   const client = await clientPromise;
@@ -33,14 +35,18 @@ export async function POST(req: Request) {
     switch (actionType) {
       case "getAllArticles": {
         const data = await collection.find().toArray();
+
         return new Response(JSON.stringify({ data }), { status: 200 });
       }
       case "comment": {
-        await collection.updateOne({_id:new BSON.ObjectId(id)},{
-          $push:{
-            history:`${authenticatedUserName} commented on ${new Date()}: ${comment}`
+        await collection.updateOne(
+          { _id: new BSON.ObjectId(id) },
+          {
+            $push: {
+              history: `${authenticatedUserName} commented on ${new Date()}: ${comment}`,
+            },
           }
-        })
+        );
         return new Response(JSON.stringify({ status: 200 }));
       }
       case "edit": {
@@ -60,6 +66,7 @@ export async function POST(req: Request) {
             },
           }
         );
+
         return new Response(JSON.stringify({ status: 200 }));
       }
       case "delete": {
@@ -67,8 +74,40 @@ export async function POST(req: Request) {
         return new Response(JSON.stringify({ status: 200 }));
       }
       case "publish": {
+        const articlesCollection: Collection<Manuscript> = await db.collection(
+          "articles"
+        );
+        const manuscriptToBePublished = await collection.findOne({
+          _id: new BSON.ObjectId(id),
+        });
+
+        if (!manuscriptToBePublished) {
+          return new Response(JSON.stringify({ data: "id aint valid" }));
+        }
+
+        await articlesCollection.insertOne(manuscriptToBePublished!);
         await collection.deleteOne({ _id: new BSON.ObjectId(id) });
-        return new Response(JSON.stringify({ status: 200 }));
+        await articlesCollection.updateOne(
+          {
+            _id: new BSON.ObjectId(id),
+          },
+          {
+            $set: {
+              date: new Date(),
+            },
+            $push: {
+              history: `${authenticatedUserName} published on ${new Date()}`,
+            },
+          }
+        );
+
+        await index.saveObject({
+          title: title,
+          author: authenticatedUserName,
+          objectID: id,
+        });
+
+        return new Response(JSON.stringify({ status: 200, data: "got it!" }));
       }
       default:
         return new Response(JSON.stringify({ error: "Invalid action type" }), {
@@ -76,6 +115,7 @@ export async function POST(req: Request) {
         });
     }
   } catch (error) {
+    console.error(error);
     return new Response(JSON.stringify({ error: "Internal Server Error" }), {
       status: 500,
     });
